chatbot-backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
python-dotenv>=1.0
groq>=0.11
pydantic>=2.10
